# api/swagger/swagger.yaml
swagger: "2.0"
...
info:
  version: "0.0.1"
  title: Education Consultant Assistant API

################################################################################
#                  Host, Base Path, Schemes and Content Types                  #
################################################################################
host: localhost:10010
...
basePath: /api/v0/
...
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json

################################################################################
#                                   Tags                                       #
################################################################################
...
parameters:
...
  customerId:
    name: customerId
    in: path
    description: The id of a customer
    type: string
    required: true
...
paths:
...
  /customers:
    x-swagger-router-controller: customers
    post:
      operationId: createCustomer
      tags:
        - Customer
      description: ""
      summary: Creates a new customer
      parameters:
        - name: newCustomer
          in: body
          description: The NewCustomer object
          required: true
          schema:
            $ref: "#/definitions/NewCustomer"
      responses:
        201:
          description: Created
          schema:
            $ref: "#/definitions/ExpandedCustomer"
        400:
          description: Bad request
          schema:
            $ref: "#/definitions/ErrorResponse"
        409:
          description: Conflict
          schema:
            $ref: "#/definitions/ErrorResponse"
        500:
          description: Internal server error
          schema:
            $ref: "#/definitions/ErrorResponse"
    get:
      operationId: getCustomers
      tags:
        - Customer
      description: ""
      summary: Returns a list of customers
      responses:
        200:
          description: OK
          schema:
            type: array
            items:
              $ref: "#/definitions/ExpandedCustomer"
        500:
          description: Internal server error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /customers/{customerId}:
    x-swagger-router-controller: customers
    parameters:
      - $ref: '#/parameters/customerId'
    get:
      operationId: getCustomerById
      tags:
        - Customer
      description: ""
      summary: Returns a customer
      responses:
        200:
          description: OK
          schema:
            $ref: "#/definitions/ExpandedCustomer"
        404:
          description: Not found
          schema:
            $ref: "#/definitions/ErrorResponse"
        500:
          description: Internal server error
          schema:
            $ref: "#/definitions/ErrorResponse"
    put:
      operationId: updateCustomer
      tags:
        - Customer
      description: ""
      summary: Updates a customer
      parameters:
        - name: updatedCustomer
          in: body
          description: The Customer object
          required: true
          schema:
            $ref: "#/definitions/Customer"
      responses:
        200:
          description: OK
          schema:
            $ref: "#/definitions/ExpandedCustomer"
        400:
          description: Bad request
          schema:
            $ref: "#/definitions/ErrorResponse"
        404:
          description: Not found
          schema:
            $ref: "#/definitions/ErrorResponse"
        500:
          description: Internal server error
          schema:
            $ref: "#/definitions/ErrorResponse"
    patch:
      operationId: partialUpdateCustomer
      tags:
        - Customer
      description: ""
      summary: Partially update a customer
      consumes:
        - application/merge-patch+json
      parameters:
        - name: customerUpdatePatch
          in: body
          description: The CustomerUpdatePatch object
          required: true
          schema:
            $ref: "#/definitions/CustomerUpdatePatch"
      responses:
        200:
          description: OK
          schema:
            $ref: "#/definitions/ExpandedCustomer"
        400:
          description: Bad request
          schema:
            $ref: "#/definitions/ErrorResponse"
        404:
          description: Not found
          schema:
            $ref: "#/definitions/ErrorResponse"
        500:
          description: Internal server error
          schema:
            $ref: "#/definitions/ErrorResponse"
    delete:
      operationId: deleteCustomer
      tags:
        - Customer
      description: ""
      summary: Deletes a customer
      responses:
        204:
          description: No content
        404:
          description: Not found
          schema:
            $ref: "#/definitions/ErrorResponse"
        500:
          description: Internal server error
          schema:
            $ref: "#/definitions/ErrorResponse"
  # /customers/{customerId}/archived:
  #   x-swagger-router-controller: customers
  #   parameters:
  #     - $ref: '#/parameters/customerId'
  #   put:
  #     operationId: updateCustomerArchived
  #     tags:
  #       - Customer
  #     description: ""
  #     summary: Updates the archived status of a customer
  #     parameters:
  #       - name: value
  #         in: query
  #         description: Set to true to archive / to false to unarchive
  #         required: true
  #         type: boolean
  #     responses:
  #       200:
  #         description: OK
  #       400:
  #         description: Bad request
  #         schema:
  #           $ref: "#/definitions/ErrorResponse"
  #       404:
  #         description: Not found
  #         schema:
  #           $ref: "#/definitions/ErrorResponse"
  #       500:
  #         description: Internal server error
  #         schema:
  #           $ref: "#/definitions/ErrorResponse"
  # /customers/{customerId}/birthday:
  #   x-swagger-router-controller: customers
  #   parameters:
  #     - $ref: '#/parameters/customerId'
  #   put:
  #     operationId: updateCustomerBirthday
  #     tags:
  #       - Customer
  #     description: ""
  #     summary: Updates the birthday of a customer
  #     parameters:
  #       - name: value
  #         in: query
  #         description: The birthday of a customer
  #         required: true
  #         type: string
  #         format: date-time
  #     responses:
  #       204:
  #         description: No content
  #       400:
  #         description: Bad request
  #         schema:
  #           $ref: "#/definitions/ErrorResponse"
  #       404:
  #         description: Not found
  #         schema:
  #           $ref: "#/definitions/ErrorResponse"
  #       500:
  #         description: Internal server error
  #         schema:
  #           $ref: "#/definitions/ErrorResponse"
  # /customers/{customerId}/visa_expiry_date:
  #   x-swagger-router-controller: customers
  #   parameters:
  #     - $ref: '#/parameters/customerId'
  #   put:
  #     operationId: updateCustomerVisaExpiryDate
  #     tags:
  #       - Customer
  #     description: ""
  #     summary: Updates the visa expiry date of a customer
  #     parameters:
  #       - name: value
  #         in: query
  #         description: The visa expiry date of a customer
  #         required: true
  #         type: string
  #         format: date-time
  #     responses:
  #       200:
  #         description: OK
  #       400:
  #         description: Bad request
  #         schema:
  #           $ref: "#/definitions/ErrorResponse"
  #       404:
  #         description: Not found
  #         schema:
  #           $ref: "#/definitions/ErrorResponse"
  #       500:
  #         description: Internal server error
  #         schema:
  #           $ref: "#/definitions/ErrorResponse"
  # /customers/{customerId}/{property}:
  #   x-swagger-router-controller: customers
  #   parameters:
  #     - $ref: '#/parameters/customerId'
  #   put:
  #     operationId: updateCustomerProperty
  #     tags:
  #       - Customer
  #     description: ""
  #     summary: Updates a customer property
  #     parameters:
  #       - name: property
  #         in: path
  #         description: The property name
  #         required: true
  #         type: string
  #         enum:
  #           - email
  #           - surname
  #           - given_name
  #           - nickname
  #           - real_name
  #           - gender
  #           - mobile
  #           - qq
  #           - wechat
  #           - au_address
  #           - foreign_address
  #       - name: value
  #         in: query
  #         description: The property value
  #         required: true
  #         type: string
  #     responses:
  #       200:
  #         description: OK
  #       400:
  #         description: Bad request
  #         schema:
  #           $ref: "#/definitions/ErrorResponse"
  #       404:
  #         description: Not found
  #         schema:
  #           $ref: "#/definitions/ErrorResponse"
  #       500:
  #         description: Internal server error
  #         schema:
  #           $ref: "#/definitions/ErrorResponse"

  ##### Checklist #####
...
definitions:
...
  NewWorkflowStatus:
    type: object
    required:
      - name
      - color
    properties:
      name:
        type: string
      color:
        $ref: "#/definitions/Color"
      pos:
        $ref: "#/definitions/SortablePosition"
    additionalProperties: false
  WorkflowStatus:
    type: object
    allOf:
      - $ref: "#/definitions/NewWorkflowStatus"
      - required:
        - _id
        - pos
        properties:
          _id:
            $ref: "#/definitions/ObjectId"
          pos:
            $ref: "#/definitions/SortablePosition"
  ExpandedWorkflowStatus:
    type: object
    allOf:
      - $ref: "#/definitions/WorkflowStatus"
      - required:
        - created_at
        - updated_at
        properties:
          created_at:
            type: string
            format: date-time
          updated_at:
            type: string
            format: date-time
...
  NewCustomer:
    type: object
    required:
      - surname
      - given_name
      - gender
      - status
    properties:
      email:
        type: string
        format: email
      surname:
        type: string
      given_name:
        type: string
      nickname:
        type: string
      real_name:
        type: string
      gender:
        $ref: "#/definitions/Gender"
      birthday:
        type: string
        format: date-time
      mobile:
        type: string
      qq:
        type: string
      wechat:
        type: string
      au_address:
        type: string
      foreign_address:
        type: string
      visa_expiry_date:
        type: string
        format: date-time
      status:
        $ref: "#/definitions/ObjectId"
      list_pos:
        $ref: "#/definitions/SortablePosition"
      workflow_pos:
        $ref: "#/definitions/SortablePosition"
      is_archived:
        type: boolean
        default: false
    additionalProperties: false
  Customer:
    type: object
    allOf:
      - $ref: "#/definitions/NewCustomer"
      - required:
          - _id
          - list_pos
          - workflow_pos
          - is_archived
        properties:
          _id:
            $ref: "#/definitions/ObjectId"
          list_pos:
            $ref: "#/definitions/SortablePosition"
          workflow_pos:
            $ref: "#/definitions/SortablePosition"
          is_archived:
            type: boolean
  ExpandedCustomer:
    type: object
    required:
      - _id
      - surname
      - given_name
      - gender
      - status
      - list_pos
      - workflow_pos
      - is_archived
      - created_at
      - updated_at
    properties:
      _id:
        $ref: "#/definitions/ObjectId"
      email:
        type: string
        format: email
      surname:
        type: string
      given_name:
        type: string
      nickname:
        type: string
      real_name:
        type: string
      gender:
        $ref: "#/definitions/Gender"
      birthday:
        type: string
        format: date-time
      mobile:
        type: string
      qq:
        type: string
      wechat:
        type: string
      au_address:
        type: string
      foreign_address:
        type: string
      visa_expiry_date:
        type: string
        format: date-time
      status:
        $ref: "#/definitions/ExpandedWorkflowStatus"
      list_pos:
        $ref: "#/definitions/SortablePosition"
      workflow_pos:
        $ref: "#/definitions/SortablePosition"
      is_archived:
        type: boolean
      checklists:
        type: array
        items:
          $ref: "#/definitions/ExpandedChecklist"
      sticky_notes:
        type: array
        items:
          $ref: "#/definitions/StickyNote"
      created_at:
        type: string
        format: date-time
      updated_at:
        type: string
        format: date-time
  CustomerUpdatePatch:
    type: object
    properties:
      email:
        type: string
        format: email
      surname:
        type: string
      given_name:
        type: string
      nickname:
        type: string
      real_name:
        type: string
      gender:
        $ref: "#/definitions/Gender"
      birthday:
        type: string
        format: date-time
      mobile:
        type: string
      qq:
        type: string
      wechat:
        type: string
      au_address:
        type: string
      foreign_address:
        type: string
      visa_expiry_date:
        type: string
        format: date-time
      status:
        $ref: "#/definitions/ObjectId"
      list_pos:
        $ref: "#/definitions/SortablePosition"
      workflow_pos:
        $ref: "#/definitions/SortablePosition"
      is_archived:
        type: boolean
    additionalProperties: false

  # Checklist
...
  ExpandedChecklist:
    type: object
    required:
      - _id
      - name
      - pos
      - created_at
      - updated_at
    properties:
      _id:
        $ref: "#/definitions/ObjectId"
      name:
        type: string
      pos:
        $ref: "#/definitions/SortablePosition"
      items:
        type: array
        items:
          $ref: "#/definitions/ExpandedChecklistItem"
      created_at:
        type: string
        format: date-time
      updated_at:
        type: string
        format: date-time
...
  NewChecklistItem:
    type: object
    required:
      - text
    properties:
      text:
        type: string
      checked:
        type: boolean
        default: false
      pos:
        $ref: "#/definitions/SortablePosition"
    additionalProperties: false
  ChecklistItem:
    type: object
    allOf:
      - $ref: "#/definitions/NewChecklistItem"
      - required:
        - _id
        - checked
        - pos
        properties:
          _id:
            $ref: "#/definitions/ObjectId"
          checked:
            type: boolean
          pos:
            $ref: "#/definitions/SortablePosition"
  ExpandedChecklistItem:
    type: object
    allOf:
      - $ref: "#/definitions/ChecklistItem"
      - required:
        - created_at
        - updated_at
        properties:
          created_at:
            type: string
            format: date-time
          updated_at:
            type: string
            format: date-time
...
  StickyNote:
    type: object
    allOf:
      - $ref: "#/definitions/NewStickyNote"
      - required:
        - _id
        - pos
        properties:
          _id:
            $ref: "#/definitions/ObjectId"
          pos:
            $ref: "#/definitions/SortablePosition"
  NewStickyNote:
    type: object
    required:
      - text
    properties:
      text:
        type: string

  # Internal types
  ObjectId:
    type: string
    pattern: ^[0-9a-fA-F]{24}$
  SortablePosition:
    type: integer
    format: int32
    minimum: 0

  # Enums
  Color:
    type: string
    enum:
      - red
      - orange
      - yellow
      - green
      - blue
      - purple
      - black
  Gender:
    type: string
    enum:
      - male
      - female

  # Responses
  ErrorResponse:
    type: object
    required:
      - statusCode
      - error
      - message
    properties:
      statusCode:
        type: integer
      error:
        type: string
      message:
        type: string
      details:
        type: array
        items:
          type: object